Conflicts:
	thegame/thegame/Game1.cs
	thegame/thegame/Player.cs

+++ thegame/thegame/MoveableEntity.cs
@@ -42,7 +42,7 @@ namespace thegame
             
         }
 
-        public void forward(int speed)
+        public void forward(float speed)
         {
             pos += look * speed;
         }
